try GCN again

+++ model/temporal/temporal_ext_gcn.py
@@ -27,7 +27,7 @@ class TemporalExtGCN(nn.Module):
         #self.gcn = GCNConv(self.node_size, self.output_size)
         self.gcn = RGCNConv(self.node_size, self.output_size, num_relations=self.num_relations)
 
-        self.fc = nn.Linear(self.node_size/2 * self.output_size, self.output_size)
+        self.fc = nn.Linear(int(self.node_size/2) * self.output_size, self.output_size)
 
         # load model parameters
         if not is_training:
